Add support for Qemu Guest Agent commands
Through Qemu we can send commands to a Qemu Guest Agent running inside
a domain.

This way we can communicate with a running Domain by asking for it's
network information, requesting a filesystem trim or even execute a
command inside a Domain.

Commands need to be send as JSON Strings, but these have been wrapped
into the QemuCommand class which has a list of pre-defined and the most
commonly used commands.

RAW commands can also be send for more flexibility.

Signed-off-by: Wido den Hollander <wido@widodh.nl>